WIP on Entity Serialization and Prefab System
