+++ GameEngine/src/Engine/Scene/Entity.h
@@ -68,7 +68,5 @@ namespace Engine
 	private:
 		entt::entity m_EntityHandle{ entt::null };
 		Scene* m_Scene = nullptr;
-
-		friend class MonoScript;
 	};
 }
