+++ GameEngine/src/Engine/Scene/EntitySerializer.cpp
@@ -1,0 +1,511 @@
+#include "enginepch.h"
+#include "Engine/Scene/EntitySerializer.h"
+
+#include "Engine/Scene/Components.h"
+#include "Engine/Scripting/ScriptEngine.h"
+#include "Engine/Physics/Physics2D.h"
+
+namespace YAML
+{
+	template<>
+	struct convert<glm::vec2>
+	{
+		static Node encode(const glm::vec2& rhs)
+		{
+			Node node;
+			node.push_back(rhs.x);
+			node.push_back(rhs.y);
+			node.SetStyle(EmitterStyle::Flow);
+			return node;
+		}
+
+		static bool decode(const Node& node, glm::vec2& rhs)
+		{
+			if (!node.IsSequence() || node.size() != 2)
+				return false;
+
+			rhs.x = node[0].as<float>();
+			rhs.y = node[1].as<float>();
+			return true;
+		}
+	};
+
+	template<>
+	struct convert<glm::vec3>
+	{
+		static Node encode(const glm::vec3& rhs)
+		{
+			Node node;
+			node.push_back(rhs.x);
+			node.push_back(rhs.y);
+			node.push_back(rhs.z);
+			node.SetStyle(EmitterStyle::Flow);
+			return node;
+		}
+
+		static bool decode(const Node& node, glm::vec3& rhs)
+		{
+			if (!node.IsSequence() || node.size() != 3)
+				return false;
+
+			rhs.x = node[0].as<float>();
+			rhs.y = node[1].as<float>();
+			rhs.z = node[2].as<float>();
+			return true;
+		}
+	};
+
+	template<>
+	struct convert<glm::vec4>
+	{
+		static Node encode(const glm::vec4& rhs)
+		{
+			Node node;
+			node.push_back(rhs.x);
+			node.push_back(rhs.y);
+			node.push_back(rhs.z);
+			node.push_back(rhs.w);
+			node.SetStyle(EmitterStyle::Flow);
+			return node;
+		}
+
+		static bool decode(const Node& node, glm::vec4& rhs)
+		{
+			if (!node.IsSequence() || node.size() != 4)
+				return false;
+
+			rhs.x = node[0].as<float>();
+			rhs.y = node[1].as<float>();
+			rhs.z = node[2].as<float>();
+			rhs.w = node[3].as<float>();
+			return true;
+		}
+	};
+
+	template<>
+	struct convert<Engine::UUID>
+	{
+		static Node encode(const Engine::UUID& uuid)
+		{
+			Node node;
+			node.push_back((uint64_t)uuid);
+			return node;
+		}
+
+		static bool decode(const Node& node, Engine::UUID& uuid)
+		{
+			uuid = node[0].as<uint64_t>();
+			return true;
+		}
+	};
+}
+
+namespace Engine
+{
+#define WRITE_SCRIPT_FIELD(FieldType, Type)			\
+	case ScriptFieldType::FieldType:				\
+		out << scriptField.GetValue<Type>();		\
+		break;										\
+
+#define READ_SCRIPT_FIELD(FieldType, Type)			\
+	case ScriptFieldType::FieldType:				\
+	{												\
+		Type data = scriptField["Data"].as<Type>();	\
+		fieldInstance.SetValue(data);				\
+		break;										\
+	}												\
+
+	YAML::Emitter& operator<<(YAML::Emitter& out, const glm::vec2& vec)
+	{
+		out << YAML::Flow;
+		out << YAML::BeginSeq << vec.x << vec.y << YAML::EndSeq;
+		return out;
+	}
+
+	YAML::Emitter& operator<<(YAML::Emitter& out, const glm::vec3& vec)
+	{
+		out << YAML::Flow;
+		out << YAML::BeginSeq << vec.x << vec.y << vec.z << YAML::EndSeq;
+		return out;
+	}
+
+	YAML::Emitter& operator<<(YAML::Emitter& out, const glm::vec4& vec)
+	{
+		out << YAML::Flow;
+		out << YAML::BeginSeq << vec.x << vec.y << vec.z << vec.w << YAML::EndSeq;
+		return out;
+	}
+
+	void EntitySerializer::Serialize(YAML::Emitter& out)
+	{
+		ENGINE_CORE_ASSERT(m_Entity.HasComponent<IDComponent>(), "Entity must have UUID!");
+
+		out << YAML::BeginMap; // Entity
+		out << YAML::Key << "Entity" << YAML::Value << m_Entity.GetUUID();
+
+		if (m_Entity.HasComponent<TagComponent>())
+		{
+			out << YAML::Key << "TagComponent";
+			out << YAML::BeginMap; // TagComponent
+
+			out << YAML::Key << "Tag" << YAML::Value << m_Entity.GetComponent<TagComponent>().Tag;
+			ENGINE_CORE_TRACE("Serializing entity with ID = {0}, name {1}", m_Entity.GetUUID(), m_Entity.GetComponent<TagComponent>().Tag);
+
+			out << YAML::EndMap; // TagComponent
+		}
+
+		if (m_Entity.HasComponent<TransformComponent>())
+		{
+			out << YAML::Key << "TransformComponent";
+			out << YAML::BeginMap; // TransformComponent
+
+			auto& transformComponent = m_Entity.GetComponent<TransformComponent>();
+			out << YAML::Key << "Position" << YAML::Value << transformComponent.Position;
+			out << YAML::Key << "Rotation" << YAML::Value << transformComponent.Rotation;
+			out << YAML::Key << "Scale" << YAML::Value << transformComponent.Scale;
+
+			out << YAML::EndMap; // TransformComponent
+		}
+
+		if (m_Entity.HasComponent<CameraComponent>())
+		{
+			out << YAML::Key << "CameraComponent";
+			out << YAML::BeginMap; // CameraComponent
+
+			auto& cameraComponent = m_Entity.GetComponent<CameraComponent>();
+			auto& camera = cameraComponent.Camera;
+
+			out << YAML::Key << "Camera";
+			out << YAML::BeginMap; // Camera
+
+			out << YAML::Key << "ProjectionType" << YAML::Value << Utils::SceneCameraProjectionTypeToString(camera.GetProjectionType());
+
+			out << YAML::Key << "PerspectiveFOV" << YAML::Value << camera.GetPerspectiveVerticalFOV();
+			out << YAML::Key << "PerspectiveNear" << YAML::Value << camera.GetPerspectiveNearClip();
+			out << YAML::Key << "PerspectiveFar" << YAML::Value << camera.GetPerspectiveFarClip();
+
+			out << YAML::Key << "OrthographicSize" << YAML::Value << camera.GetOrthographicSize();
+			out << YAML::Key << "OrthographicNear" << YAML::Value << camera.GetOrthographicNearClip();
+			out << YAML::Key << "OrthographicFar" << YAML::Value << camera.GetOrthographicFarClip();
+
+			out << YAML::EndMap; // Camera
+
+			out << YAML::Key << "Primary" << YAML::Value << cameraComponent.Primary;
+			out << YAML::Key << "FixedAspectRatio" << YAML::Value << cameraComponent.FixedAspectRatio;
+
+			out << YAML::EndMap; // CameraComponent
+		}
+
+		if (m_Entity.HasComponent<SpriteRendererComponent>())
+		{
+			out << YAML::Key << "SpriteRendererComponent";
+			out << YAML::BeginMap; // SpriteRendererComponent
+
+			auto& spriteRendererComponent = m_Entity.GetComponent<SpriteRendererComponent>();
+			out << YAML::Key << "Color" << YAML::Value << spriteRendererComponent.Color;
+			out << YAML::Key << "Path" << YAML::Value << spriteRendererComponent.Path.string();
+			out << YAML::Key << "Tiling" << YAML::Value << spriteRendererComponent.Tiling;
+
+			out << YAML::Key << "IsSubTexture" << YAML::Value << spriteRendererComponent.IsSubTexture;
+			out << YAML::Key << "SubCoords" << YAML::Value << spriteRendererComponent.SubCoords;
+			out << YAML::Key << "SubCellSize" << YAML::Value << spriteRendererComponent.SubCellSize;
+			out << YAML::Key << "SubSpriteSize" << YAML::Value << spriteRendererComponent.SubSpriteSize;
+
+			out << YAML::EndMap; // SpriteRendererComponent
+		}
+
+		if (m_Entity.HasComponent<CircleRendererComponent>())
+		{
+			out << YAML::Key << "CircleRendererComponent";
+			out << YAML::BeginMap; // CircleRendererComponent
+
+			auto& circleRendererComponent = m_Entity.GetComponent<CircleRendererComponent>();
+			out << YAML::Key << "Color" << YAML::Value << circleRendererComponent.Color;
+			out << YAML::Key << "Radius" << YAML::Value << circleRendererComponent.Radius;
+			out << YAML::Key << "Thickness" << YAML::Value << circleRendererComponent.Thickness;
+			out << YAML::Key << "Fade" << YAML::Value << circleRendererComponent.Fade;
+
+			out << YAML::EndMap; // CircleRendererComponent
+		}
+
+		if (m_Entity.HasComponent<ScriptComponent>())
+		{
+			out << YAML::Key << "ScriptComponent";
+			out << YAML::BeginMap; // ScriptComponent
+
+			auto& scriptComponent = m_Entity.GetComponent<ScriptComponent>();
+			out << YAML::Key << "ClassName" << YAML::Value << scriptComponent.ClassName;
+
+			// Fields
+			const auto& scriptFields = ScriptEngine::GetEntityClasses().at(scriptComponent.ClassName)->GetScriptFields();
+			if (scriptFields.size() > 0)
+			{
+				out << YAML::Key << "ScriptFields" << YAML::Value;
+				auto& entityFields = ScriptEngine::GetScriptFieldMap(m_Entity);
+				out << YAML::BeginSeq;
+				for (auto const& [name, field] : scriptFields)
+				{
+					if (entityFields.find(name) == entityFields.end())
+						continue;
+
+					out << YAML::BeginMap; // ScriptField
+					out << YAML::Key << "Name" << YAML::Value << name;
+					out << YAML::Key << "Type" << YAML::Value << Utils::ScriptFieldTypeToString(field.Type);
+
+					out << YAML::Key << "Data" << YAML::Value;
+					ScriptFieldInstance& scriptField = entityFields.at(name);
+
+					switch (field.Type)
+					{
+						WRITE_SCRIPT_FIELD(Float, float);
+						WRITE_SCRIPT_FIELD(Double, double);
+						WRITE_SCRIPT_FIELD(Bool, bool);
+						WRITE_SCRIPT_FIELD(Char, char);
+						WRITE_SCRIPT_FIELD(String, std::string);
+						WRITE_SCRIPT_FIELD(SByte, int8_t);
+						WRITE_SCRIPT_FIELD(Short, int16_t);
+						WRITE_SCRIPT_FIELD(Int, int32_t);
+						WRITE_SCRIPT_FIELD(Long, int64_t);
+						WRITE_SCRIPT_FIELD(Byte, uint16_t); // upcasting uint8_t to unint16_t to fix yaml-cpp encode/decode
+						WRITE_SCRIPT_FIELD(UShort, uint16_t);
+						WRITE_SCRIPT_FIELD(UInt, uint32_t);
+						WRITE_SCRIPT_FIELD(ULong, uint64_t);
+						WRITE_SCRIPT_FIELD(Vector2, glm::vec2);
+						WRITE_SCRIPT_FIELD(Vector3, glm::vec3);
+						WRITE_SCRIPT_FIELD(Vector4, glm::vec4);
+						WRITE_SCRIPT_FIELD(Entity, UUID);
+					default:
+						ENGINE_CORE_ERROR("Script Field Type {} does not support serialization!", Utils::ScriptFieldTypeToString(field.Type));
+					}
+
+					out << YAML::EndMap; // ScriptFields
+				}
+
+				out << YAML::EndSeq;
+			}
+
+			out << YAML::EndMap; // ScriptComponent
+		}
+
+		if (m_Entity.HasComponent<Rigidbody2DComponent>())
+		{
+			out << YAML::Key << "Rigidbody2DComponent";
+			out << YAML::BeginMap; // Rigidbody2DComponent
+
+			auto& rigidbody2DComponent = m_Entity.GetComponent<Rigidbody2DComponent>();
+			out << YAML::Key << "Type" << YAML::Value << Utils::Rigidbody2DBodyTypeToString(rigidbody2DComponent.Type);
+			out << YAML::Key << "FixedRotation" << YAML::Value << rigidbody2DComponent.FixedRotation;
+			out << YAML::Key << "Smoothing" << YAML::Value << Utils::Rigidbody2DSmoothingTypeToString(rigidbody2DComponent.Smoothing);
+
+			out << YAML::EndMap; // Rigidbody2DComponent
+		}
+
+		if (m_Entity.HasComponent<BoxCollider2DComponent>())
+		{
+			out << YAML::Key << "BoxCollider2DComponent";
+			out << YAML::BeginMap; // BoxCollider2DComponent
+
+			auto& boxCollider2DComponent = m_Entity.GetComponent<BoxCollider2DComponent>();
+			out << YAML::Key << "Offset" << YAML::Value << boxCollider2DComponent.Offset;
+			out << YAML::Key << "Size" << YAML::Value << boxCollider2DComponent.Size;
+			out << YAML::Key << "Density" << YAML::Value << boxCollider2DComponent.Density;
+			out << YAML::Key << "Friction" << YAML::Value << boxCollider2DComponent.Friction;
+			out << YAML::Key << "Restitution" << YAML::Value << boxCollider2DComponent.Restitution;
+			out << YAML::Key << "RestitutionThreshold" << YAML::Value << boxCollider2DComponent.RestitutionThreshold;
+
+			out << YAML::EndMap; // BoxCollider2DComponent
+		}
+
+		if (m_Entity.HasComponent<CircleCollider2DComponent>())
+		{
+			out << YAML::Key << "CircleCollider2DComponent";
+			out << YAML::BeginMap; // CircleCollider2DComponent
+
+			auto& circleCollider2DComponent = m_Entity.GetComponent<CircleCollider2DComponent>();
+			out << YAML::Key << "Offset" << YAML::Value << circleCollider2DComponent.Offset;
+			out << YAML::Key << "Radius" << YAML::Value << circleCollider2DComponent.Radius;
+			out << YAML::Key << "Density" << YAML::Value << circleCollider2DComponent.Density;
+			out << YAML::Key << "Friction" << YAML::Value << circleCollider2DComponent.Friction;
+			out << YAML::Key << "Restitution" << YAML::Value << circleCollider2DComponent.Restitution;
+			out << YAML::Key << "RestitutionThreshold" << YAML::Value << circleCollider2DComponent.RestitutionThreshold;
+
+			out << YAML::EndMap; // CircleCollider2DComponent
+		}
+
+		out << YAML::EndMap; // Entity
+	}
+
+	Engine::Entity EntitySerializer::Deserialize(YAML::Node& entity)
+	{
+		UUID uuid = entity["Entity"].as<uint64_t>();
+
+		std::string name;
+		auto tagComponent = entity["TagComponent"];
+		if (tagComponent)
+			name = tagComponent["Tag"].as<std::string>();
+
+		ENGINE_CORE_TRACE("Deserializing entity {0}: {1}", uuid, name);
+		m_Entity = m_Scene->CreateEntityWithUUID(uuid, name);
+
+		auto transformComponent = entity["TransformComponent"];
+		if (transformComponent)
+		{
+			// Entities are created with a transform
+			auto& tc = m_Entity.GetComponent<TransformComponent>();
+			tc.Position = transformComponent["Position"].as<glm::vec3>();
+			tc.Rotation = transformComponent["Rotation"].as<glm::vec3>();
+			tc.Scale = transformComponent["Scale"].as<glm::vec3>();
+		}
+		else // So if there is none then remove from Entity
+		{
+			m_Entity.RemoveComponent<TransformComponent>();
+		}
+
+		auto cameraComponent = entity["CameraComponent"];
+		if (cameraComponent)
+		{
+			auto& cc = m_Entity.AddComponent<CameraComponent>();
+			auto& camera = cc.Camera;
+
+			auto& cameraProps = cameraComponent["Camera"];
+
+			camera.SetProjectionType(Utils::SceneCameraProjectionTypeFromString(cameraProps["ProjectionType"].as<std::string>()));
+
+			camera.SetPerspectiveVerticalFOV(cameraProps["PerspectiveFOV"].as<float>());
+			camera.SetPerspectiveNearClip(cameraProps["PerspectiveNear"].as<float>());
+			camera.SetPerspectiveFarClip(cameraProps["PerspectiveFar"].as<float>());
+
+			camera.SetOrthographicSize(cameraProps["OrthographicSize"].as<float>());
+			camera.SetOrthographicNearClip(cameraProps["OrthographicNear"].as<float>());
+			camera.SetOrthographicFarClip(cameraProps["OrthographicFar"].as<float>());
+
+			cc.Primary = cameraComponent["Primary"].as<bool>();
+			cc.FixedAspectRatio = cameraComponent["FixedAspectRatio"].as<bool>();
+		}
+
+		auto spriteRendererComponent = entity["SpriteRendererComponent"];
+		if (spriteRendererComponent)
+		{
+			auto& spriteRenderer = m_Entity.AddComponent<SpriteRendererComponent>();
+			spriteRenderer.Color = spriteRendererComponent["Color"].as<glm::vec4>();
+			spriteRenderer.Path = spriteRendererComponent["Path"].as<std::string>();
+			spriteRenderer.Tiling = spriteRendererComponent["Tiling"].as<float>();
+
+			spriteRenderer.IsSubTexture = spriteRendererComponent["IsSubTexture"].as<bool>();
+			spriteRenderer.SubCoords = spriteRendererComponent["SubCoords"].as<glm::vec2>();
+			spriteRenderer.SubCellSize = spriteRendererComponent["SubCellSize"].as<glm::vec2>();
+			spriteRenderer.SubSpriteSize = spriteRendererComponent["SubSpriteSize"].as<glm::vec2>();
+
+			spriteRenderer.LoadTexture(spriteRenderer.Path);
+		}
+
+		auto circleRendererComponent = entity["CircleRendererComponent"];
+		if (circleRendererComponent)
+		{
+			auto& circleRenderer = m_Entity.AddComponent<CircleRendererComponent>();
+			circleRenderer.Color = circleRendererComponent["Color"].as<glm::vec4>();
+			circleRenderer.Radius = circleRendererComponent["Radius"].as<float>();
+			circleRenderer.Thickness = circleRendererComponent["Thickness"].as<float>();
+			circleRenderer.Fade = circleRendererComponent["Fade"].as<float>();
+		}
+
+		auto scriptComponent = entity["ScriptComponent"];
+		if (scriptComponent)
+		{
+			auto& sc = m_Entity.AddComponent<ScriptComponent>();
+			sc.ClassName = scriptComponent["ClassName"].as<std::string>();
+
+			auto scriptFields = scriptComponent["ScriptFields"];
+			if (scriptFields)
+			{
+				Ref<ScriptClass> entityClass = ScriptEngine::GetEntityClass(sc.ClassName);
+				if (entityClass)
+				{
+					const auto& fields = entityClass->GetScriptFields();
+					auto& entityFields = ScriptEngine::GetScriptFieldMap(m_Entity);
+
+					for (auto scriptField : scriptFields)
+					{
+						std::string name = scriptField["Name"].as<std::string>();
+						std::string typeString = scriptField["Type"].as<std::string>();
+						ScriptFieldType type = Utils::ScriptFieldTypeFromString(typeString);
+
+						ScriptFieldInstance& fieldInstance = entityFields[name];
+
+						//ENGINE_CORE_ASSERT(fields.find(name) != fields.end()); // extra field in saved file
+
+						if (fields.find(name) == fields.end())
+							continue;
+
+						fieldInstance.Field = fields.at(name);
+
+						switch (type)
+						{
+							READ_SCRIPT_FIELD(Float, float);
+							READ_SCRIPT_FIELD(Double, double);
+							READ_SCRIPT_FIELD(Bool, bool);
+							READ_SCRIPT_FIELD(Char, char);
+							READ_SCRIPT_FIELD(String, std::string);
+							READ_SCRIPT_FIELD(SByte, int8_t);
+							READ_SCRIPT_FIELD(Short, int16_t);
+							READ_SCRIPT_FIELD(Int, int32_t);
+							READ_SCRIPT_FIELD(Long, int64_t);
+							READ_SCRIPT_FIELD(Byte, uint16_t); // upcasting uint8_t to unint16_t to fix yaml-cpp encode/decode
+							READ_SCRIPT_FIELD(UShort, uint16_t);
+							READ_SCRIPT_FIELD(UInt, uint32_t);
+							READ_SCRIPT_FIELD(ULong, uint64_t);
+							READ_SCRIPT_FIELD(Vector2, glm::vec2);
+							READ_SCRIPT_FIELD(Vector3, glm::vec3);
+							READ_SCRIPT_FIELD(Vector4, glm::vec4);
+							READ_SCRIPT_FIELD(Entity, UUID);
+						default:
+							ENGINE_CORE_ERROR("Script Field Type {} does not support deserialization!", Utils::ScriptFieldTypeToString(type));
+						}
+					}
+				}
+				else
+				{
+					ENGINE_CORE_ERROR("Entity Class does not exists for given script component!");
+				}
+			}
+		}
+
+		auto rigidbody2DComponent = entity["Rigidbody2DComponent"];
+		if (rigidbody2DComponent)
+		{
+			auto& rigidbody2D = m_Entity.AddComponent<Rigidbody2DComponent>();
+			rigidbody2D.Type = Utils::Rigidbody2DBodyTypeFromString(rigidbody2DComponent["Type"].as<std::string>());
+			rigidbody2D.FixedRotation = rigidbody2DComponent["FixedRotation"].as<bool>();
+			rigidbody2D.Smoothing = Utils::Rigidbody2DSmoothingTypeFromString(rigidbody2DComponent["Smoothing"].as<std::string>());
+		}
+
+		auto boxCollider2DComponent = entity["BoxCollider2DComponent"];
+		if (boxCollider2DComponent)
+		{
+			auto& boxCollider2D = m_Entity.AddComponent<BoxCollider2DComponent>();
+			boxCollider2D.Offset = boxCollider2DComponent["Offset"].as<glm::vec2>();
+			boxCollider2D.Size = boxCollider2DComponent["Size"].as<glm::vec2>();
+
+			boxCollider2D.Density = boxCollider2DComponent["Density"].as<float>();
+			boxCollider2D.Friction = boxCollider2DComponent["Friction"].as<float>();
+			boxCollider2D.Restitution = boxCollider2DComponent["Restitution"].as<float>();
+			boxCollider2D.RestitutionThreshold = boxCollider2DComponent["RestitutionThreshold"].as<float>();
+		}
+
+		auto circleCollider2DComponent = entity["CircleCollider2DComponent"];
+		if (circleCollider2DComponent)
+		{
+			auto& circleCollider2D = m_Entity.AddComponent<CircleCollider2DComponent>();
+			circleCollider2D.Offset = circleCollider2DComponent["Offset"].as<glm::vec2>();
+			circleCollider2D.Radius = circleCollider2DComponent["Radius"].as<float>();
+
+			circleCollider2D.Density = circleCollider2DComponent["Density"].as<float>();
+			circleCollider2D.Friction = circleCollider2DComponent["Friction"].as<float>();
+			circleCollider2D.Restitution = circleCollider2DComponent["Restitution"].as<float>();
+			circleCollider2D.RestitutionThreshold = circleCollider2DComponent["RestitutionThreshold"].as<float>();
+		}
+
+		return m_Entity;
+	}
+
+}
